Removed search bar from simplecontactview, so that the user can use their own implementation, filter method provided
