+++ app/src/main/java/com/luisdc05/simplecontactpickerdemo/MainActivity.kt
@@ -1,8 +1,13 @@
 package com.luisdc05.simplecontactpickerdemo
 
+import android.annotation.SuppressLint
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
+import android.support.design.widget.TextInputEditText
+import android.text.Editable
+import android.text.TextWatcher
 import android.util.Log
+import android.view.MotionEvent
 import android.widget.Button
 import com.luisdc05.simplecontactpicker.PickedContacts
 import com.luisdc05.simplecontactpicker.SimpleContactPicker
@@ -15,7 +20,9 @@ class MainActivity : AppCompatActivity(), OnContactsReceived, ContactSelectionLi
 
     private lateinit var contactPicker: SimpleContactPicker
     private lateinit var pickedContacts: PickedContacts
+    private lateinit var searchInput: TextInputEditText
 
+    @SuppressLint("ClickableViewAccessibility")
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
@@ -34,6 +41,33 @@ class MainActivity : AppCompatActivity(), OnContactsReceived, ContactSelectionLi
         button.setOnClickListener {
             Log.d(TAG, contactPicker.selectedContacts.size.toString())
         }
+
+        searchInput = findViewById(R.id.search)
+        showSearchIcon()
+        searchInput.setOnTouchListener { view, motionEvent ->
+            val DRAWABLE_RIGHT = 2
+            if (motionEvent.action == MotionEvent.ACTION_UP) {
+                if (motionEvent.rawX >= (searchInput.right - searchInput.compoundDrawables[DRAWABLE_RIGHT].getBounds().width())) {
+                    searchInput.setText("")
+                }
+            }
+            false
+        }
+        searchInput.addTextChangedListener(object: TextWatcher {
+            override fun afterTextChanged(p0: Editable?) {
+                val text = searchInput.text.toString()
+                if (text == "") {
+                    showSearchIcon()
+                } else {
+                    showClearIcon()
+                }
+                contactPicker.filter(text)
+            }
+            override fun beforeTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {
+            }
+            override fun onTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {
+            }
+        })
     }
 
     override fun onResume() {
@@ -58,5 +92,13 @@ class MainActivity : AppCompatActivity(), OnContactsReceived, ContactSelectionLi
 
     override fun beforeSelection(contact: ContactBase): Boolean {
         return true
+    }
+
+    private fun showClearIcon() {
+        searchInput.setCompoundDrawablesWithIntrinsicBounds(0, 0, R.drawable.ic_delete, 0)
+    }
+
+    private fun showSearchIcon() {
+        searchInput.setCompoundDrawablesWithIntrinsicBounds(0, 0, R.drawable.ic_search, 0)
     }
 }
