+++ simplecontactpicker/build.gradle
@@ -29,7 +29,6 @@ dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 
     implementation "com.android.support:recyclerview-v7:$android_support_version"
-    implementation "com.android.support:design:$android_support_version"
     implementation "com.squareup.picasso:picasso:2.5.2"
     implementation "jp.wasabeef:picasso-transformations:2.1.2"
     testImplementation 'junit:junit:4.12'
